+++ backend/src/controllers/commerce/orders_controller.js
@@ -16,18 +16,6 @@ export async function addOrder(req, res) {
   }
 }
 
-export async function updateOrder(req, res) {
-  try {
-    const id = req.params.id;
-    const { status } = req.body;
-    const order = await orderService.updateOrderService({ id, status });
-    res.status(200).json(order);
-  } catch (error) {
-    console.error("Error updating order", error);
-    res.status(500).json({ message: "Internal Server Error" });
-  }
-}
-
 export async function getCustomerOrders(req, res) {
   try {
     const userId = req.user.id;
