+++ backend/src/services/admin/commerce/admin_order_services.js
@@ -1,0 +1,11 @@
+import { Order } from "../../../model/index.js";
+
+export async function updateOrderService(order) {
+  const { id, status } = order;
+  const updateOrder = await Order.findOneAndUpdate(
+    { orderCode: id },
+    { status: status },
+    { new: true },
+  );
+  return updateOrder;
+}
