+++ backend/src/services/commerce/order_services.js
@@ -47,13 +47,3 @@ export async function addOrderService(customer) {
 
   return savedOrder;
 }
-
-export async function updateOrderService(order) {
-  const { id, status } = order;
-  const updateOrder = await Order.findOneAndUpdate(
-    { orderCode: id },
-    { status: status },
-    { new: true },
-  );
-  return updateOrder;
-}
